pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "black>=22.12.0",
    "pylint>=2.15.10",
]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "hxse", email = "narutohy@163.com"},
]
dependencies = [
    "hid>=1.0.5",
    "fire>=0.5.0",
    "pywin32>=305",
    "psutil>=5.9.4",
    "rich>=13.0.1",
    "pyserial>=3.5",
]
requires-python = ">=3.10"
license = {text = "MIT"}
